requirements.txt
# Napoleon Cybersecurity AI - Production Requirements
# Machine Learning Core
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0

# AI/LLM Integration
litellm[proxy]>=1.40.0
backoff>=2.2.1
httpx>=0.26.0
asyncio-mqtt>=0.16.0
anthropic>=0.25.0
openai>=1.12.0

# Web Framework
fastapi>=0.108.0
uvicorn[standard]>=0.27.0
aiohttp>=3.9.0
python-multipart>=0.0.6

# CLI and UI
rich>=13.7.0
click>=8.1.0
prompt-toolkit>=3.0.43

# Data Processing
pyyaml>=6.0.1
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
jinja2>=3.1.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.9.0

# Security and Networking
cryptography>=41.0.0
paramiko>=3.4.0
scapy>=2.5.0

# Template Engine
mako>=1.3.0

# MCP Integration  
mcp>=1.0.0

# Development and Utilities
setuptools>=69.0.0
wheel>=0.42.0
pip>=24.0
